doneee yayyyy

+++ DeepData.Steganograph/Interfaces/IProgress.cs
@@ -2,5 +2,5 @@ namespace DeepData.Interfaces;
 
 public interface IProgress
 {
-    void Update(int current, int total);
+    void Update(int current, long total);
 }
